api/requirements.txt
requests>=2.27.1
flask-redis==0.4.0
sendgrid>=6.9.6
python-dotenv>=0.19.2
flask>=2.1.0
autoscraper>=1.1.14
transformers>=4.12.0
flask-cors>=3.0.0
passlib>=1.7.4
authlib>=1.0.0
gunicorn>=20.1.0
